+++ transektcount/src/main/java/com/wmstein/transektcount/AddSpeciesActivity.java
@@ -39,7 +39,7 @@ import androidx.core.app.NavUtils;
  * by 3 interdependent and correlated arrays in arrays.xml
  * <p>
  * Created for TourCount by wmstein on 2019-04-12,
- * last edited on 2022-04-30
+ * last edited on 2023-05-08
  */
 public class AddSpeciesActivity extends AppCompatActivity implements SharedPreferences.OnSharedPreferenceChangeListener
 {