+++ transektcount/src/main/java/com/wmstein/transektcount/CountingLActivity.java
@@ -58,14 +58,14 @@ import androidx.appcompat.app.AppCompatActivity;
  * landscape orientation.
  * The reason for adding this activity is the limitation of the used drop-down spinner 
  * which produces a crash for a null object reference when forced to use another orientation.
- * 
+ * <p>
  * CountingLActivity does the actual counting on landscape layout with 12 counters, 
  * checks for alerts, calls CountOptionsLActivity, EditSectionLActivity and DummyLActivity, clones a section,
  * switches screen off when device is pocketed and lets you send a message.
- * 
+ * <p>
  * Inspired by milo's CountingActivity.java of BeeCount from 05/05/2014.
  * Changes and additions for TransektCount by wmstein since 18.02.2016
- * Last edit on 2020-05-02
+ * Last edit on 2023-05-08
  */
 public class CountingLActivity extends AppCompatActivity implements SharedPreferences.OnSharedPreferenceChangeListener
 {
@@ -580,7 +580,7 @@ public class CountingLActivity extends AppCompatActivity implements SharedPrefer
     /************************
      * The functions below are triggered by the count buttons
      * and righthand/lefthand (LH) views
-     *
+     <p>
      * countUpf1i is triggered by buttonUpf1i in widget_counting_i.xml
      */
     public void countUpf1i(View view)
@@ -1777,9 +1777,9 @@ public class CountingLActivity extends AppCompatActivity implements SharedPrefer
     /**
      * Following functions are taken from the Apache commons-lang3-3.4 library
      * licensed under Apache License Version 2.0, January 2004
-     * 
+     <p> 
      * Checks if a CharSequence is whitespace, empty ("") or null
-     * 
+      <p> 
      * isBlank(null)      = true
      * isBlank("")        = true
      * isBlank(" ")       = true
@@ -1808,7 +1808,7 @@ public class CountingLActivity extends AppCompatActivity implements SharedPrefer
 
     /**
      * Checks if a CharSequence is not empty (""), not null and not whitespace only.
-     * 
+      <p> 
      * isNotBlank(null)      = false
      * isNotBlank("")        = false
      * isNotBlank(" ")       = false